feat: move tags to tag doctype
