+++ frappe/desk/doctype/tag/tag.json
@@ -5,16 +5,24 @@
  "editable_grid": 1,
  "engine": "InnoDB",
  "field_order": [
-  "description"
+  "description",
+  "count"
  ],
  "fields": [
   {
    "fieldname": "description",
    "fieldtype": "Small Text",
    "label": "Description"
+  },
+  {
+   "default": "0",
+   "fieldname": "count",
+   "fieldtype": "Int",
+   "label": "Count",
+   "read_only": 1
   }
  ],
- "modified": "2019-09-16 22:40:02.027484",
+ "modified": "2019-09-17 00:57:31.675904",
  "modified_by": "Administrator",
  "module": "Desk",
  "name": "Tag",
