+++ frappe/public/js/frappe/ui/tag_editor.js
@@ -35,7 +35,7 @@ frappe.ui.TagEditor = Class.extend({
 			onTagAdd: (tag) => {
 				if(me.initialized && !me.refreshing) {
 					return frappe.call({
-						method: 'frappe.desk.tags.add_tag',
+						method: 'frappe.desk.doctype.tag.tag.add_tag',
 						args: me.get_args(tag),
 						callback: function(r) {
 							var user_tags = me.user_tags ? me.user_tags.split(",") : [];
@@ -49,7 +49,7 @@ frappe.ui.TagEditor = Class.extend({
 			onTagRemove: (tag) => {
 				if(!me.refreshing) {
 					return frappe.call({
-						method: 'frappe.desk.tags.remove_tag',
+						method: 'frappe.desk.doctype.tag.tag.remove_tag',
 						args: me.get_args(tag),
 						callback: function(r) {
 							var user_tags = me.user_tags.split(",");
@@ -82,7 +82,7 @@ frappe.ui.TagEditor = Class.extend({
 		$input.on("input", function(e) {
 			var value = e.target.value;
 			frappe.call({
-				method:"frappe.desk.tags.get_tags",
+				method:"frappe.desk.doctype.tag.tag.get_tags",
 				args:{
 					doctype: me.frm.doctype,
 					txt: value.toLowerCase(),
