+++ frappe/utils/global_tags.py
@@ -13,7 +13,7 @@ from frappe.utils import cint, strip_html_tags
 from frappe.model.base_document import get_controller
 from six import text_type
 
-def setup_global_search_table():
+def setup_global_tags_table():
 	"""
 	Creates __global_search table
 	:return:
@@ -28,11 +28,14 @@ def reset():
 	"""
 	frappe.db.sql('DELETE FROM `__global_tags`')
 
-def delete_global_tags_records_for_document(doctype, docname):
+def delete_global_tags_for_document(doc):
+	"""
+		Deletes doc entry from __global_tags
+	"""
 	frappe.db.sql("""DELETE
 		FROM `__global_search`
 		WHERE doctype = %s
-			AND name = %s""", (doctype, docname, tag))
+			AND name = %s""", (doc.doctype, doc.name))
 
 
 def update_global_tags(doc, tags):
@@ -98,10 +101,6 @@ def get_documents_for_tag(tag):
 		}, as_dict=True)
 
 	for res in result:
-		# if res.doctype in results.keys():
-		# 	results[res.doctype].append([res])
-		# else:
-		# 	results[res.doctype] = [[res]]
 		if res.doctype in results.keys():
 			results[res.doctype].append(res)
 		else:
